Generalize from Okta to just SAML
In some cases, this also means specifically accommodating Google.

+++ cmd/substrate-create-admin-account/idp.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"log"
+	"strings"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials/stscreds"
@@ -19,10 +20,11 @@ import (
 const (
 	Administrator = "Administrator"
 	Auditor       = "Auditor"
+	Google        = "Google"
 	Okta          = "Okta"
 )
 
-func okta(sess *session.Session, account *organizations.Account, metadata string) {
+func idp(sess *session.Session, account *organizations.Account, metadata string) {
 	org, err := awsorgs.DescribeOrganization(organizations.New(sess))
 	if err != nil {
 		log.Fatal(err)
@@ -35,8 +37,15 @@ func okta(sess *session.Session, account *organizations.Account, metadata string
 		)),
 	})
 
-	ui.Spin("configuring Okta as your organization's identity provider")
-	saml, err := awsiam.EnsureSAMLProvider(svc, Okta, metadata)
+	name := "IdP"
+	if strings.Contains(metadata, "google.com") {
+		name = Google
+	} else if strings.Contains(metadata, "okta.com") {
+		name = Okta
+	}
+
+	ui.Spinf("configuring %s as your organization's identity provider", name)
+	saml, err := awsiam.EnsureSAMLProvider(svc, name, metadata)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -44,16 +53,18 @@ func okta(sess *session.Session, account *organizations.Account, metadata string
 	//log.Printf("%+v", saml)
 
 	// Give Okta some entrypoints in the Admin account.
-	ui.Spin("finding or creating roles for Okta to use in the ops account")
+	ui.Spinf("finding or creating roles for %s to use in this admin account", name)
 	assumeRolePolicyDocument := &policies.Document{
 		Statement: []policies.Statement{
 			policies.AssumeRolePolicyDocument(&policies.Principal{
 				AWS: []string{
 					aws.StringValue(org.MasterAccountId),
-					roles.Arn(
-						aws.StringValue(account.Id),
-						"substrate-credential-factory",
-					),
+					/*
+						roles.Arn(
+							aws.StringValue(account.Id),
+							"substrate-credential-factory",
+						),
+					*/
 					users.Arn(
 						aws.StringValue(org.MasterAccountId),
 						users.OrganizationAdministrator,
@@ -124,44 +135,59 @@ func okta(sess *session.Session, account *organizations.Account, metadata string
 	}
 	ui.Stop("ok")
 
-	// And give Okta a user that can enumerate the roles it can assume.
-	ui.Spin("finding or creating a user for Okta to use to enumerate roles")
-	user, err := awsiam.EnsureUserWithPolicy(
-		svc,
-		Okta,
-		&policies.Document{
-			Statement: []policies.Statement{
-				policies.Statement{
-					Action:   []string{"iam:ListAccountAliases", "iam:ListRoles"},
-					Resource: []string{"*"},
+	// Google asks GSuite admins to set custom attributes user by user.  Help
+	// these poor souls out by at least telling them exactly what value to set.
+	if name == Google {
+		ui.Printf(
+			`set the AWS/Role custom attribute in GSuite for every authorized AWS console user to "%s,%s"`,
+			roles.Arn(aws.StringValue(account.Id), Administrator),
+			saml.Arn,
+		)
+		if _, err := ui.Prompt("press <enter> after you've configured at least one GSuite user (so you can test this)"); err != nil {
+			log.Fatal(err)
+		}
+	}
+
+	// Give Okta a user that can enumerate the roles it can assume.  Only Okta
+	// needs this.  Google puts more of the burden on GSuite admins.
+	if name == Okta {
+		ui.Spin("finding or creating a user for Okta to use to enumerate roles")
+		user, err := awsiam.EnsureUserWithPolicy(
+			svc,
+			Okta,
+			&policies.Document{
+				Statement: []policies.Statement{
+					policies.Statement{
+						Action:   []string{"iam:ListAccountAliases", "iam:ListRoles"},
+						Resource: []string{"*"},
+					},
 				},
 			},
-		},
-	)
-	if err != nil {
-		log.Fatal(err)
-	}
-	ui.Stopf("user %s", user.UserName)
-	//log.Printf("%+v", user)
-	if ok, err := ui.Confirm("do you need to configure Okta's AWS integration? (yes or no)"); err != nil {
-		log.Fatal(err)
-	} else if ok {
-		ui.Spin("deleting existing access keys and creating a new one")
-		if err := awsiam.DeleteAllAccessKeys(svc, Okta); err != nil {
-			log.Fatal(err)
-		}
-		accessKey, err := awsiam.CreateAccessKey(svc, aws.StringValue(user.UserName))
+		)
 		if err != nil {
 			log.Fatal(err)
 		}
-		ui.Stop("ok")
-		//log.Printf("%+v", accessKey)
-		ui.Printf("Okta needs this SAML provider ARN: %s", saml.Arn)
-		ui.Printf(".. and this access key ID: %s", accessKey.AccessKeyId)
-		ui.Printf("...and this secret access key: %s", accessKey.SecretAccessKey)
-		_, err = ui.Prompt("press ENTER after you've updated your Okta configuration")
-		if err != nil {
+		ui.Stopf("user %s", user.UserName)
+		//log.Printf("%+v", user)
+		if ok, err := ui.Confirm("do you need to configure Okta's AWS integration? (yes or no)"); err != nil {
 			log.Fatal(err)
+		} else if ok {
+			ui.Spin("deleting existing access keys and creating a new one")
+			if err := awsiam.DeleteAllAccessKeys(svc, Okta); err != nil {
+				log.Fatal(err)
+			}
+			accessKey, err := awsiam.CreateAccessKey(svc, aws.StringValue(user.UserName))
+			if err != nil {
+				log.Fatal(err)
+			}
+			ui.Stop("ok")
+			//log.Printf("%+v", accessKey)
+			ui.Printf("Okta needs this SAML provider ARN: %s", saml.Arn)
+			ui.Printf(".. and this access key ID: %s", accessKey.AccessKeyId)
+			ui.Printf("...and this secret access key: %s", accessKey.SecretAccessKey)
+			if _, err := ui.Prompt("press <enter> after you've updated your Okta configuration"); err != nil {
+				log.Fatal(err)
+			}
 		}
 	}
